requirements.txt
Authlib==0.15.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
cryptography==36.0.0
dataclasses-json==0.5.6
idna==3.3
marshmallow==3.14.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
oauthlib==3.1.1
pycparser==2.21
python-decouple==3.5
requests==2.26.0
requests-oauthlib==1.3.0
tweepy==4.4.0
typing-inspect==0.7.1
typing_extensions==4.0.1
urllib3==1.26.7

psycopg~=3.0.5
mysql-connector-python~=8.0.27